requirements.txt
streamlit
google-generativeai
toml